+++ src/index/index_writer.rs
@@ -17,17 +17,17 @@
 pub struct IndexWriter;
 /// Maximum number of documents. We subtract 128 to ensure it's well below the typical JVM's
 /// `ArrayUtil.MAX_ARRAY_LENGTH` and avoid potential overflow issues across JVM implementations.
-pub const MAX_DOCS: u32 = (i32::MAX - 128) as u32;
+pub const MAX_DOCS: i32 = i32::MAX - 128;
 
 /// Maximum value for the token position in an indexed field.
-pub const MAX_POSITION: u32 = (i32::MAX - 128) as u32;
+pub const MAX_POSITION: i32 = i32::MAX - 128;
 
 /// A variable that holds the actual maximum number of documents, which can be adjusted for testing purposes.
-pub const ACTUAL_MAX_DOCS: u32 = MAX_DOCS;
+pub const ACTUAL_MAX_DOCS: i32 = MAX_DOCS;
 impl IndexWriter {
     pub fn set_live_commit_data(&self) {}
 
-    pub fn get_actual_max_docs() -> u32 {
+    pub fn get_actual_max_docs() -> i32 {
         ACTUAL_MAX_DOCS
     }
 }
